Веб программирование/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
requests==2.31.0
email-validator==2.3.0
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
fastapi-sso==0.10.0
redis==5.0.1

